+++ spec/lib/sidekiq/pauzer_spec.rb
@@ -20,13 +20,13 @@ RSpec.describe Sidekiq::Pauzer do
   describe ".pause!" do
     it "adds queue to the paused list" do
       expect { %w[foo bar].each { |q| described_class.pause!(q) } }
-        .to change { redis_smembers(described_class.redis_key) }.to(match_array(%w[foo bar]))
+        .to change { redis_smembers(described_class::REDIS_KEY) }.to(match_array(%w[foo bar]))
         .and change(described_class, :paused_queues).to(match_array(%w[foo bar]))
     end
 
     it "support queue name given as Symbol" do
       expect { %i[foo bar].each { |q| described_class.pause!(q) } }
-        .to change { redis_smembers(described_class.redis_key) }.to(match_array(%w[foo bar]))
+        .to change { redis_smembers(described_class::REDIS_KEY) }.to(match_array(%w[foo bar]))
         .and change(described_class, :paused_queues).to(match_array(%w[foo bar]))
     end
 
@@ -34,7 +34,7 @@ RSpec.describe Sidekiq::Pauzer do
       described_class.pause!("foo")
 
       expect { %w[foo bar].each { |q| described_class.pause!(q) } }
-        .to change { redis_smembers(described_class.redis_key) }.to(match_array(%w[foo bar]))
+        .to change { redis_smembers(described_class::REDIS_KEY) }.to(match_array(%w[foo bar]))
         .and change(described_class, :paused_queues).to(match_array(%w[foo bar]))
     end
   end
@@ -47,19 +47,19 @@ RSpec.describe Sidekiq::Pauzer do
 
     it "removes queue from the paused list" do
       expect { described_class.unpause!("foo") }
-        .to change { redis_smembers(described_class.redis_key) }.to(contain_exactly("bar"))
+        .to change { redis_smembers(described_class::REDIS_KEY) }.to(contain_exactly("bar"))
         .and change(described_class, :paused_queues).to(contain_exactly("bar"))
     end
 
     it "support queue name given as Symbol" do
       expect { described_class.unpause!(:foo) }
-        .to change { redis_smembers(described_class.redis_key) }.to(contain_exactly("bar"))
+        .to change { redis_smembers(described_class::REDIS_KEY) }.to(contain_exactly("bar"))
         .and change(described_class, :paused_queues).to(contain_exactly("bar"))
     end
 
     it "skips non-paused queues" do
       expect { described_class.unpause!("baz") }
-        .to keep_unchanged { redis_smembers(described_class.redis_key) }
+        .to keep_unchanged { redis_smembers(described_class::REDIS_KEY) }
         .and keep_unchanged(described_class, :paused_queues)
     end
   end
@@ -101,13 +101,10 @@ RSpec.describe Sidekiq::Pauzer do
     end
 
     it "allows re-entrance" do
-      described_class.configure do |c|
-        c.key_prefix   = "foobar:"
-        c.refresh_rate = 42
-      end
+      described_class.configure { |c| c.refresh_rate = 42.0 }
 
       expect { |b| described_class.configure(&b) }
-        .to yield_with_args(have_attributes(key_prefix: "foobar:", refresh_rate: 42))
+        .to yield_with_args(have_attributes(refresh_rate: 42.0))
     end
   end